Extract and cleanup list-delete tests
